Use conditional compilation for incompatible targets
