+++ src/lib.rs
@@ -1,113 +1,32 @@
 extern crate regex;
 
-use std::process::Command;
-mod lsb_release;
-mod rhel_release;
-mod sw_vers;
-mod utils;
+#[cfg(target_os = "android")]
+#[path = "android/mod.rs"]
+mod imp;
 
-///A list of supported operating system types
-#[derive(Debug)]
-#[derive(PartialEq)]
-#[derive(Clone)]
-pub enum OSType {
-    Unknown,
-    Redhat,
-    OSX,
-    Ubuntu,
-    Debian,
-    Arch,
-    CentOS,
-}
+#[cfg(target_os = "emscripten")]
+#[path = "emscripten/mod.rs"]
+mod imp;
 
-/// Holds information about Operating System type and its version
-/// If the version could not be fetched it defaults to `0.0.0`
-#[derive(Debug)]
-#[derive(Clone)]
-#[derive(PartialEq)]
-pub struct OSInformation {
-    pub os_type: self::OSType,
-    pub version: String,
-}
+#[cfg(target_os = "linux")]
+#[path = "linux/mod.rs"]
+mod imp;
 
-fn default_version() -> String {
-    "0.0.0".into()
-}
+#[cfg(target_os = "macos")]
+#[path = "macos/mod.rs"]
+mod imp;
 
-fn unknown_os() -> OSInformation {
-    OSInformation {
-        os_type: OSType::Unknown,
-        version: default_version(),
-    }
-}
+#[cfg(target_os = "redox")]
+#[path = "redox/mod.rs"]
+mod imp;
 
-fn is_os_x() -> bool {
-    match Command::new("sw_vers").output() {
-        Ok(output) => output.status.success(),
-        Err(_) => false,
-    }
-}
+#[cfg(windows)]
+#[path = "windows/mod.rs"]
+mod imp;
 
-fn get_sw_vers() -> OSInformation {
-    if let Some(osx_info) = sw_vers::retrieve() {
-        OSInformation {
-            os_type: OSType::OSX,
-            version: osx_info.product_version.unwrap_or_else(default_version),
-        }
-    } else {
-        unknown_os()
-    }
-}
+mod os_info;
 
-fn lsb_release() -> OSInformation {
-    match lsb_release::retrieve() {
-        Some(release) => {
-            if release.distro == Some("Ubuntu".to_string()) {
-                OSInformation {
-                    os_type: OSType::Ubuntu,
-                    version: release.version.unwrap_or_else(default_version),
-                }
-            } else if release.distro == Some("Debian".to_string()) {
-                OSInformation {
-                    os_type: OSType::Debian,
-                    version: release.version.unwrap_or_else(default_version),
-                }
-            } else if release.distro == Some("Arch".to_string()) {
-                OSInformation {
-                    os_type: OSType::Arch,
-                    version: release.version.unwrap_or_else(default_version),
-                }
-            } else if release.distro == Some("CentOS".to_string()) {
-                OSInformation {
-                    os_type: OSType::CentOS,
-                    version: release.version.unwrap_or_else(default_version),
-                }
-            } else {
-                unknown_os()
-            }
-        }
-        None => unknown_os(),
-    }
-}
-
-fn rhel_release() -> OSInformation {
-    match rhel_release::retrieve() {
-        Some(release) => {
-            if release.distro == Some("CentOS".to_string()) {
-                OSInformation {
-                    os_type: OSType::CentOS,
-                    version: release.version.unwrap_or_else(default_version),
-                }
-            } else {
-                OSInformation {
-                    os_type: OSType::Redhat,
-                    version: release.version.unwrap_or_else(default_version),
-                }
-            }
-        }
-        None => unknown_os(),
-    }
-}
+use os_info::OSInformation;
 
 ///Returns the current operating system type
 ///
@@ -120,15 +39,5 @@ fn rhel_release() -> OSInformation {
 ///println!("Version: {}", os.version);
 ///```
 pub fn current_platform() -> OSInformation {
-    if is_os_x() {
-        get_sw_vers()
-    } else if lsb_release::is_available() {
-        lsb_release()
-    } else if utils::file_exists("/etc/redhat-release") ||
-               utils::file_exists("/etc/centos-release")
-    {
-        rhel_release()
-    } else {
-        unknown_os()
-    }
+    imp::current_platform()
 }
